last:

+++ database/migrations/2020_12_14_122821_create_obat_resep_table.php
@@ -23,7 +23,7 @@ class CreateObatResepTable extends Migration
             $table->enum('waktu_minum', ['Sebelum Makan', 'Saat Makan','Sesudah Makan']);
             $table->enum('keterangan', ['Kondisional', 'Harus Habis', 'Rutin']);
             $table->string('jml_obat');
-            $table->integer('jml_kali_minum');
+            $table->float('jml_kali_minum');
             $table->timestamps();
 
             $table->foreign('id_obat')->references('id')->on('obat')->onDelete('cascade');
